refactor: 이벤트 콜백함수 모듈화
